Randomize juju model names
Our test was hiding the fact that spark spells need an updated lxd profile to
handle some of the libkmod stuff in zookeeper and others. Because it was running
last and re-using existing juju models that had lxd profiles applied b/c of cdk
and openstack running prior we were hiding the fact that the spark spells would
actually fail if run alone.

ref: https://github.com/conjure-up/spells/issues/130

Signed-off-by: Adam Stokes <battlemidget@users.noreply.github.com>